+++ jgclark.Dashboard/src/react/components/ItemNoteLink.jsx
@@ -1,14 +1,13 @@
 // @flow
 //--------------------------------------------------------------------------
 // Dashboard React component to show Note Links after main item content
-// Last updated 2025-11-13 for v2.3.0 by @jgclark
+// Last updated 2025-11-16 for v2.3.0 by @jgclark
 //--------------------------------------------------------------------------
 import React from 'react'
 import type { TSection, TSectionItem } from '../../types.js'
 import { useAppContext } from './AppContext.jsx'
 import TooltipOnKeyPress from './ToolTipOnModifierPress.jsx'
 import { isDailyDateStr, isWeeklyDateStr, isMonthlyDateStr, isQuarterlyDateStr } from '@helpers/dateTime'
-import { getFolderFromFilename } from '@helpers/folders'
 import { parseTeamspaceFilename, TEAMSPACE_FA_ICON } from '@helpers/teamspace'
 import { logDebug, clo } from '@helpers/react/reactDev'
 import { extractModifierKeys } from '@helpers/react/reactMouseKeyboard.js'
@@ -30,16 +29,13 @@ function ItemNoteLink({ item, thisSection, alwaysShowNoteTitle = false }: Props)
 
   // ------ COMPUTED VALUES --------------------------------
 
-  const { sendActionToPlugin, dashboardSettings, reactSettings } = useAppContext()
+  const { sendActionToPlugin, reactSettings } = useAppContext()
   const filename = item.para?.filename ?? '<no filename found>'
   // compute the things we need later
   const noteTitle = item?.para?.title || ''
   // logDebug(`ItemNoteLink`, `ItemNoteLink for item.itemFilename:${filename} noteTitle:${noteTitle} item.para.noteType:${item.para.noteType} / thisSection.filename=${thisSection.sectionFilename}`)
-  const folderNamePart = dashboardSettings.showFolderName && filename !== '<no filename found>' && getFolderFromFilename(filename) !== '/' ? `${getFolderFromFilename(filename)} /` : ''
 
   // Compute the icon and link class and style depending whether this is a teamspace item, and/or note types
-  // const isFromTeamspace = item.para?.isFromTeamspace ?? false
-
   const noteIconToUse = (isDailyDateStr(filename))
     ? 'fa-regular fa-calendar-star'
     : (isWeeklyDateStr(filename))
@@ -50,21 +46,20 @@ function ItemNoteLink({ item, thisSection, alwaysShowNoteTitle = false }: Props)
           ? 'fa-regular fa-calendar-range'
           : 'fa-light fa-file-lines'
   const parsedTeamspace = parseTeamspaceFilename(filename)
-  const filenameWithoutTeamspacePrefix = parsedTeamspace.filename
   const isFromTeamspace = parsedTeamspace.isTeamspace
+  const filenameWithoutTeamspacePrefix = parsedTeamspace.filename
+  // FIXME: this is still returning '/' for Teamspace calendar notes
+  const folderNamePart = (parsedTeamspace.filepath !== '/') ? `${parsedTeamspace.filepath} /` : ''
   // logDebug(`ItemNoteLink`, `noteIconToUse:${noteIconToUse} with filenameWithoutTeamspacePrefix:${filenameWithoutTeamspacePrefix}`)
-  // const linkClass = 'noteTitle'
-  // const linkStyle = 'folderName'
   const showNoteTitle = alwaysShowNoteTitle || item.para?.noteType === 'Notes' || filenameWithoutTeamspacePrefix !== thisSection.sectionFilename
 
-  // TODO: Break this out into a separate component
+  // Show Teamspace indicator and name, if this is a Teamspace note
   let teamspaceIndicator = null
   if (isFromTeamspace) {
     const teamspaceTitle = item.teamspaceTitle && item.teamspaceTitle !== 'Unknown Teamspace' ? item.teamspaceTitle : ''
     teamspaceIndicator = (
-      <span className='pad-left teamspaceName pad-right'>
-        <i className={`${TEAMSPACE_FA_ICON} pad-right`}></i>
-        {teamspaceTitle}
+      <span className='pad-left teamspaceName pad-right-larger'>
+        <i className={`${TEAMSPACE_FA_ICON} pad-right`}></i>{teamspaceTitle}
       </span>
     )
   }