+++ jgclark.Dashboard/CHANGELOG.md
@@ -10,7 +10,7 @@ For more details see the [plugin's documentation](https://github.com/NotePlan/pl
 - TODO: enlarged default window size on iPads
 -->
 
-## [2.3.0.b14] 2025-10-13
+## [2.3.0.b14] 2025-10-16
 - Can now change the displayed order of Sections, using a new panel in the Settings window. This is set per Perspective (if used). It can be reset to the default sort order.
 - Edit dialog improvements:
   - The text box (where the task text can be ed) is now focused by default
@@ -18,6 +18,8 @@ For more details see the [plugin's documentation](https://github.com/NotePlan/pl
   - Further polish to layout at specific widths
 - Fix to double-escaped JSON in settings.json file. Updated all places it writes, and added a `parseSettings` to tidy up existing files.
 - Fix for isModified setting for tags (#703, for @dwertheimer)
+- Fix for styling of note links that are from Teamspace notes
+- ??? Fix for missing referenced paras
 
 ## [2.3.0.b13] 2025-10-11???
 - When lower-priority items filter is turned on, change the move "All → ..." buttons to show "All shown → ..." instead.